// types/satdik.ts
export type Satdik = {
  RowID: number
  Tanggal_Data: string
  Is_Manual_Update: number
  Nama: string
  Jenjang: string
  Alamat: string
  Pimpinan: string
  Website: string
  Jumlah_Peserta_Didik: number
  Jumlah_Alumni: number
  Jumlah_Pendidik: number
  Jumlah_Tenaga_Kependidikan: number
  Lokasi_Lintang: string
  Lokasi_Bujur: string
  Akreditasi_Lembaga: string
  Akreditasi_Prodi: string
  Kode_Satker: string
}